sync patch r10863, r10396, r10864, r10873 from main trunk. Add security check to make code run safe.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/branches/UDK2010@11554 6f19259b-4bc3-4df7-8a09-765794883524

+++ EdkCompatibilityPkg/Foundation/Library/Dxe/EfiIfrSupportLib/IfrOpCodeCreation.c
@@ -51,7 +51,7 @@ Returns:
   EFI_IFR_SUBTITLE  Subtitle;
 
   Subtitle.Header.OpCode  = EFI_IFR_SUBTITLE_OP;
-  Subtitle.Header.Length  = sizeof (EFI_IFR_SUBTITLE);
+  Subtitle.Header.Length  = (UINT8) sizeof (EFI_IFR_SUBTITLE);
   Subtitle.SubTitle       = StringToken;
 
   EfiCopyMem (FormBuffer, &Subtitle, sizeof (EFI_IFR_SUBTITLE));
@@ -100,7 +100,7 @@ Returns:
   EFI_IFR_TEXT  Text;
 
   Text.Header.OpCode  = EFI_IFR_TEXT_OP;
-  Text.Header.Length  = sizeof (EFI_IFR_TEXT);
+  Text.Header.Length  = (UINT8) sizeof (EFI_IFR_TEXT);
   Text.Text           = StringToken;
 
   Text.TextTwo        = StringTokenTwo;
@@ -155,7 +155,7 @@ Returns:
   EFI_IFR_REF Hyperlink;
 
   Hyperlink.Header.OpCode = EFI_IFR_REF_OP;
-  Hyperlink.Header.Length = sizeof (EFI_IFR_REF);
+  Hyperlink.Header.Length = (UINT8) sizeof (EFI_IFR_REF);
   Hyperlink.FormId        = FormId;
   Hyperlink.Prompt        = StringToken;
   Hyperlink.Help          = StringTokenTwo;
@@ -228,7 +228,7 @@ Returns:
   }
 
   OneOf.Header.OpCode = EFI_IFR_ONE_OF_OP;
-  OneOf.Header.Length = sizeof (EFI_IFR_ONE_OF);
+  OneOf.Header.Length = (UINT8) sizeof (EFI_IFR_ONE_OF);
   OneOf.QuestionId    = QuestionId;
   OneOf.Width         = DataWidth;
   OneOf.Prompt        = PromptToken;
@@ -243,7 +243,7 @@ Returns:
 
   for (Index = 0; Index < OptionCount; Index++) {
     OneOfOption.Header.OpCode = EFI_IFR_ONE_OF_OPTION_OP;
-    OneOfOption.Header.Length = sizeof (EFI_IFR_ONE_OF_OPTION);
+    OneOfOption.Header.Length = (UINT8) sizeof (EFI_IFR_ONE_OF_OPTION);
 
     OneOfOption.Option        = OptionsList[Index].StringToken;
     OneOfOption.Value         = OptionsList[Index].Value;
@@ -255,7 +255,7 @@ Returns:
     LocalBuffer = (UINT8 *) (LocalBuffer + sizeof (EFI_IFR_ONE_OF_OPTION));
   }
 
-  EndOneOf.Header.Length  = sizeof (EFI_IFR_END_ONE_OF);
+  EndOneOf.Header.Length  = (UINT8) sizeof (EFI_IFR_END_ONE_OF);
   EndOneOf.Header.OpCode  = EFI_IFR_END_ONE_OF_OP;
 
   EfiCopyMem (LocalBuffer, &EndOneOf, sizeof (EFI_IFR_END_ONE_OF));
@@ -316,7 +316,7 @@ Returns:
   UINT8                 *LocalBuffer;
 
   OrderedList.Header.OpCode = EFI_IFR_ORDERED_LIST_OP;
-  OrderedList.Header.Length = sizeof (EFI_IFR_ORDERED_LIST);
+  OrderedList.Header.Length = (UINT8) sizeof (EFI_IFR_ORDERED_LIST);
   OrderedList.QuestionId    = QuestionId;
   OrderedList.MaxEntries    = MaxEntries;
   OrderedList.Prompt        = PromptToken;
@@ -331,7 +331,7 @@ Returns:
 
   for (Index = 0; Index < OptionCount; Index++) {
     OrderedListOption.Header.OpCode = EFI_IFR_ONE_OF_OPTION_OP;
-    OrderedListOption.Header.Length = sizeof (EFI_IFR_ONE_OF_OPTION);
+    OrderedListOption.Header.Length = (UINT8) sizeof (EFI_IFR_ONE_OF_OPTION);
 
     OrderedListOption.Option        = OptionsList[Index].StringToken;
     OrderedListOption.Value         = OptionsList[Index].Value;
@@ -343,7 +343,7 @@ Returns:
     LocalBuffer = (UINT8 *) (LocalBuffer + sizeof (EFI_IFR_ONE_OF_OPTION));
   }
 
-  EndOrderedList.Header.Length  = sizeof (EFI_IFR_END_ONE_OF);
+  EndOrderedList.Header.Length  = (UINT8) sizeof (EFI_IFR_END_ONE_OF);
   EndOrderedList.Header.OpCode  = EFI_IFR_END_ONE_OF_OP;
 
   EfiCopyMem (LocalBuffer, &EndOrderedList, sizeof (EFI_IFR_END_ONE_OF));
@@ -406,7 +406,7 @@ Returns:
   }
 
   CheckBox.Header.OpCode  = EFI_IFR_CHECKBOX_OP;
-  CheckBox.Header.Length  = sizeof (EFI_IFR_CHECK_BOX);
+  CheckBox.Header.Length  = (UINT8) sizeof (EFI_IFR_CHECK_BOX);
   CheckBox.QuestionId     = QuestionId;
   CheckBox.Width          = DataWidth;
   CheckBox.Prompt         = PromptToken;
@@ -486,7 +486,7 @@ Returns:
   }
 
   Numeric.Header.OpCode = EFI_IFR_NUMERIC_OP;
-  Numeric.Header.Length = sizeof (EFI_IFR_NUMERIC);
+  Numeric.Header.Length = (UINT8) sizeof (EFI_IFR_NUMERIC);
   Numeric.QuestionId    = QuestionId;
   Numeric.Width         = DataWidth;
   Numeric.Prompt        = PromptToken;
@@ -555,7 +555,7 @@ Returns:
   EFI_IFR_STRING  String;
 
   String.Header.OpCode  = EFI_IFR_STRING_OP;
-  String.Header.Length  = sizeof (EFI_IFR_STRING);
+  String.Header.Length  = (UINT8) sizeof (EFI_IFR_STRING);
   String.QuestionId     = QuestionId;
   String.Width          = DataWidth;
   String.Prompt         = PromptToken;
@@ -604,7 +604,7 @@ Returns:
   EFI_IFR_BANNER  Banner;
 
   Banner.Header.OpCode  = EFI_IFR_BANNER_OP;
-  Banner.Header.Length  = sizeof (EFI_IFR_BANNER);
+  Banner.Header.Length  = (UINT8) sizeof (EFI_IFR_BANNER);
   EfiCopyMem (&Banner.Title, &Title, sizeof (UINT16));
   EfiCopyMem (&Banner.LineNumber, &LineNumber, sizeof (UINT16));
   Banner.Alignment = Alignment;
